+++ src/main.jsx
@@ -1,7 +1,8 @@
 import { StrictMode, useEffect, useState } from 'react'
 import { createRoot } from 'react-dom/client'
-import { createBrowserRouter, Navigate, RouterProvider, useNavigate } from 'react-router-dom'
-import { getTeste } from "./requests.ts"
+import { createBrowserRouter, Navigate, RouterProvider, useNavigate, Link, Outlet } from 'react-router-dom'
+
+//import { getTeste } from "./requests.js"
 
 import SignUp_Up from './pages/SignUp_In/Up/Up.jsx'
 import SignUp_In from './pages/SignUp_In/In/in.jsx'
@@ -10,18 +11,19 @@ import Clientes from './pages/Cadastro/Clientes.jsx'
 import Fornecedores from './pages/Fornecedores/Fornecedores.jsx'
 import Pedidos from './pages/Pedidos/Pedidos.jsx'
 import Usuarios from './pages/Usuarios/Usuarios.jsx'
-import SideBar from './SideBar.jsx'
 
 import './index.css'
+import List from './example/List.jsx'
+import Teste from './example/Teste.jsx'
 
 function Auth({ children }) {
     const navigate = useNavigate();
     const [auth, setAuth] = useState(true);
 
     useEffect(() => {
-        getTeste().then((value) => {
-            setAuth(value);
-        })
+        // getTeste().then((value) => {
+        //     setAuth(value);
+        // })
     }, []);
 
     useEffect(() => {
@@ -33,6 +35,25 @@ function Auth({ children }) {
 
 function Redirec() {
     return <Navigate to="/login" />;
+}
+
+function SideBar() {
+    return (
+        <>
+            <div className="sidebar">
+                <div className="sidebar-header">ProductFolio</div>
+                <ul className="sidebar-menu">
+                    <li><Link to="/">Dashboard</Link></li>
+                    <li><Link to="/produtos">Produtos</Link></li>
+                    <li><Link to="/clientes">Clientes</Link></li>
+                    <li><Link to="/fornecedores">Fornecedores</Link></li>
+                    <li><Link to="/pedidos">Pedidos</Link></li>
+                    <li><Link to="/usuarios">Usuários</Link></li>
+                </ul>
+            </div>
+            <Outlet/>
+        </>
+    );
 }
 
 const router = createBrowserRouter([
@@ -47,6 +68,14 @@ const router = createBrowserRouter([
     {
         path: "/signin",
         element: <SignUp_In />
+    },
+	{
+        path: "/example/list",
+        element: <List />
+    },
+    {
+        path: "/example/teste",
+        element: <Teste />
     },
     {
         path: "/",
